pipelineClaude/requirements.txt
langfuse==3.6.2
langchain==0.3.25
langchain-openai==0.3.14
python-dotenv
ulid-py
pandas
numpy
matplotlib
networkx
scikit-learn
